feat: Implement a wide range of updates and fixes
This commit introduces a variety of new features, bug fixes, and performance improvements across the application.

- **Footer Data:** The Footer component now dynamically displays the address and phone number from the API, with fallbacks to default values.
- **Safari Style Fix:** A bug where the ItemCard's background effect did not correctly revert its style in Safari has been fixed.
- **Android TabList Fix:** An issue where the active tab in the TabList did not scroll into view on Android devices has been resolved.
- **Routing:** The main page content has been moved to the `/menu` route, and a redirect has been added from the root (`/`).
- **Performance:** The scroll event handler has been further optimized by increasing the throttle delay to improve performance on fast scrolls.
- **Flavor Logic:** The logic for displaying flavor options has been updated to use exact matching for item names.
- **Dynamic Titles:** The application now displays dynamic item titles based on flavor selections (e.g., "Mocha").
- **Price Formatting:** The price formatting logic has been improved to handle more complex price strings.
