+++ src/app/Menu/page.tsx
@@ -5,6 +5,7 @@ import Header from "@/components/Header";
 import Footer from "@/components/Footer";
 import { throttle } from "../utils/throttle";
 import MenuStyle from "./page.style";
+import { LinearProgress } from "@mui/material";
 import TabList from "@/components/TabList";
 import CategorySection from "@/components/CategorySection";
 import useRestaurantProfile from "@/app/utils/useRestaurantProfile";
@@ -17,6 +18,7 @@ import supportsEmoji from "../utils/SupportsEmoji";
 const DEFAULT_RESTAURANT_ID = "c7f3a9e2-1b4d-4f8e-9a6c-7d2e3b9f1c84";
 
 export default function MenuPage() {
+  const [isReady, setIsReady] = useState(false);
   const [activeCategory, setActiveCategory] = useState<string>("");
   const categoryRefs = useRef<{[key: string]: HTMLDivElement | null}>({});
   const tabListRef = useRef<HTMLDivElement | null>(null);
@@ -74,18 +76,63 @@ useEffect(() => {
     }
   };
 
-  const throttledScrollHandler = throttle(handleScroll, 200);
+  const throttledScrollHandler = throttle(handleScroll, 300);
 
   window.addEventListener('scroll', throttledScrollHandler);
   return () => window.removeEventListener('scroll', throttledScrollHandler);
 }, [finalMenuData, activeCategory]);
 
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setIsReady(true);
+    }, 800); // adjust delay as needed
+
+    return () => clearTimeout(timer);
+  }, []);
 
   
   return (
-    <MenuStyle>
-      <Header logoId={restaurantData?.data?.logoIds?.[0]} />
-      <div ref={tabListRef}>
+    <div style={{ position: "relative" }}>
+      {/* Loading screen */}
+      {!isReady && (
+        <div
+          style={{
+            position: "fixed",
+            top: 0,
+            left: 0,
+            width: "100vw",
+            height: "100vh",
+            backgroundColor: "#0284c7", // blue background
+            zIndex: 9999,
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "flex-start",
+            alignItems: "center",
+          }}
+        >
+          <LinearProgress
+            sx={{
+              width: "100%",
+              height: "4px",
+              backgroundColor: "#60a5fa", // lighter blue track
+              "& .MuiLinearProgress-bar": {
+                backgroundColor: "#ffffff", // white progress bar
+              },
+            }}
+          />
+        </div>
+      )}
+
+      {/* Main content */}
+      <div
+        style={{
+          opacity: isReady ? 1 : 0,
+          transition: "opacity 0.4s ease-in-out",
+        }}
+      >
+        <MenuStyle>
+          <Header logoId={restaurantData?.data?.logoIds?.[0]} />
+          <div ref={tabListRef}>
         <TabList 
           categories={finalMenuData?.data || []} 
           activeCategory={activeCategory}
@@ -106,7 +153,7 @@ useEffect(() => {
           />
         </div>
       ))}
-      <Footer address={restaurantData?.data?.address} />
+      <Footer address={restaurantData?.data?.address} phone={restaurantData?.data?.phone} />
       <ScrollToTopButton />
       <div style={{ textAlign: "center", marginTop: "24px", marginBottom: "16px" }}>
         <span
@@ -127,5 +174,7 @@ useEffect(() => {
         </div>
 
     </MenuStyle>
+      </div>
+    </div>
   );
 }
