+++ src/components/CategorySection.style.ts
@@ -112,11 +112,17 @@ export const ItemCard = styled.div<{ isActive?: boolean }>`
 }
 
 ${({ isActive }) =>
-  isActive &&
-  `
+  isActive
+    ? `
     &::before {
       opacity: 0.6;
       filter: hue-rotate(45deg) brightness(1.2);
+    }
+  `
+    : `
+    &::before {
+      opacity: 0.9;
+      filter: drop-shadow(0 4px 6px rgba(0, 0, 0, 0.2));
     }
   `}
